requirements.txt
fastapi
uvicorn
anthropic
chromadb
pypdf
python-dotenv
python-multipart
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
aiofiles